pyproject.toml
[tool.poetry]
name = "blog"
version = "0.0.1"
description = ""
authors = ["zengxiaolou <zengevent@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
drf-yasg = "^1.20.0"
djangorestframework = "^3.12.4"
loguru = "^0.5.3"
django-filter = "^2.4.0"
celery = "^5.1.2"
PyMySQL = "^1.0.2"
django-cors-headers = "^3.7.0"
social-auth-app-django = "^4.0.0"
djangorestframework-jwt = "^1.11.0"
redis = "^3.5.3"
django-redis = "^5.0.0"
tencentcloud-sdk-python = "^3.0.452"
qiniu = "^7.4.1"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
